+++ src/lingua/extract.py
@@ -52,8 +52,9 @@ class POEntry(polib.POEntry):
     def tcomment(self, value):
         pass
 
-    def update(self, message):
-        self.occurrences.append(message.location)
+    def update(self, message, add_occurrences=True):
+        if add_occurrences:
+            self.occurrences.append(message.location)
         self.flags.extend(f for f in message.flags if f not in self.flags)
         if message.comment not in self._comments:
             self._comments.append(message.comment)
@@ -199,6 +200,9 @@ def main():
     parser.add_argument('--no-location',
             action='store_false', dest='location',
             help='Do not include location information')
+    parser.add_argument('--no-linenumbers',
+            action='store_false', dest='nolinenumbers',
+            help='Replace linenumbers with -1 and only include each file once to reduce differences')
     parser.add_argument('-n', '--add-location',
             action='store_true', dest='location', default=True,
             help='Include location information (default)')
@@ -272,7 +276,7 @@ def main():
                     entry.msgstr_plural[0] = ''
                     entry.msgstr_plural[1] = ''
                 catalog.append(entry)
-            entry.update(message)
+            entry.update(message, add_occurrences=options.location)
         scanned += 1
     if not scanned:
         print('No files scanned, aborting', file=sys.stderr)
@@ -286,6 +290,17 @@ def main():
     elif options.sort == 'location':
         # Order the occurrences themselves, so the output is consistent
         catalog.sort(key=lambda m: m.occurrences.sort() or m.occurrences)
+
+    if not options.nolinenumbers:
+        for entry in catalog:
+            seen = set()
+            occurrences = []
+            for location, line in entry.occurrences:
+                if location in seen:
+                    continue
+                occurrences.append((location, -1))
+                seen.add(location)
+            entry.occurrences = occurrences
 
     catalog.save(options.output)