+++ changes.rst
@@ -13,6 +13,10 @@ Changelog
 - Use domain filtering for Python expressions where possible (i.e. when using
   translaction functions that take a domain parameter, such as ``dgettext``.
 
+- Fix --no-location option to actually work.
+
+- Add --no-linenumbers option.
+
 
 3.12 - October 10, 2015
 -----------------------